requirements.txt
pillow==5.4.1
numpy==1.16.1
pandas==0.20.3
scipy==0.19.1
scikit-learn==0.19.1
scikit-image==0.14.2
rawpy==0.10.1
Glymur==0.8.17
python-dateutil==2.8.0
https://download.pytorch.org/whl/cu80/torch-1.0.1.post2-cp35-cp35m-linux_x86_64.whl
torchvision==0.2.2
